feat(messages work and reroute)
