+++ server/views/pages/pop/appointments.njk
@@ -5,7 +5,7 @@
 {%- from "govuk/components/button/macro.njk" import govukButton %}
 
 
-{% set pageTitle = applicationName + " - Home" %}
+{% set pageTitle = applicationName + " - Appointments" %}
 {% set mainClasses = "app-container govuk-body" %}
 
 {% block content %}
