+++ server/views/pages/pop/conditions.njk
@@ -4,7 +4,7 @@
 
 {% from "govuk/components/warning-text/macro.njk" import govukWarningText %}
 
-{% set pageTitle = applicationName + " - License" %}
+{% set pageTitle = applicationName + " - Your conditions" %}
 {% set mainClasses = "app-container govuk-body" %}
 
 {% block content %}
