+++ server/views/pages/pop/details.njk
@@ -1,6 +1,6 @@
 {% extends "../../partials/layout.njk" %}
 
-{% set pageTitle = applicationName + " - Home" %}
+{% set pageTitle = applicationName + " - Your details" %}
 {% set mainClasses = "app-container govuk-body" %}
 
 {% block content %}
