+++ server/views/pages/pop/index.njk
@@ -37,28 +37,35 @@
         <li class="gem-c-cards__list-item">
           <div class="gem-c-cards__list-item-wrapper">
             <h3 class="gem-c-cards__sub-heading govuk-heading-s">
-              <a class="govuk-link govuk-link--no-underline gem-c-cards__link" href="/your-details">Your details</a>
+              <a class="govuk-link govuk-link--no-underline gem-c-cards__link" href="/pop/your-details">Your details</a>
             </h3>
           </div>
         </li>
         <li class="gem-c-cards__list-item">
           <div class="gem-c-cards__list-item-wrapper">
             <h3 class="gem-c-cards__sub-heading govuk-heading-s">
-              <a class="govuk-link govuk-link--no-underline gem-c-cards__link" href="/your-progress">Your progress</a>
+              <a class="govuk-link govuk-link--no-underline gem-c-cards__link" href="/pop/your-progress">Your progress</a>
             </h3>
           </div>
         </li>
         <li class="gem-c-cards__list-item">
           <div class="gem-c-cards__list-item-wrapper">
             <h3 class="gem-c-cards__sub-heading govuk-heading-s">
-              <a class="govuk-link govuk-link--no-underline gem-c-cards__link" href="/appointments">Past and upcoming appointments</a>
+              <a class="govuk-link govuk-link--no-underline gem-c-cards__link" href="/pop/appointments">Past and upcoming appointments</a>
+            </h3>
+          </div>
+        </li>
+         <li class="gem-c-cards__list-item">
+          <div class="gem-c-cards__list-item-wrapper">
+            <h3 class="gem-c-cards__sub-heading govuk-heading-s">
+              <a class="govuk-link govuk-link--no-underline gem-c-cards__link" href="/pop/messages">Messages</a>
             </h3>
           </div>
         </li>
         <li class="gem-c-cards__list-item">
           <div class="gem-c-cards__list-item-wrapper">
             <h3 class="gem-c-cards__sub-heading govuk-heading-s">
-              <a class="govuk-link govuk-link--no-underline gem-c-cards__link" href="/conditions">Your conditions</a>
+              <a class="govuk-link govuk-link--no-underline gem-c-cards__link" href="pop/conditions">Your conditions</a>
             </h3>
           </div>
         </li>
