+++ server/views/pages/pop/new-message.njk
@@ -1,0 +1,63 @@
+{% extends "../../partials/layout.njk" %}
+
+{%- from "moj/components/ticket-panel/macro.njk" import mojTicketPanel -%}
+{% from "govuk/components/warning-text/macro.njk" import govukWarningText %}
+{%- from "govuk/components/button/macro.njk" import govukButton %}
+{% from "govuk/components/tag/macro.njk" import govukTag %}
+
+{% set pageTitle = applicationName + " - Create new message" %}
+{% set mainClasses = "app-container govuk-body" %}
+
+{% block content %}
+
+<div class="govuk-grid-row">
+  <div class="govuk-grid-column-full">
+    <div class="govuk-grid-column govuk-!-margin-top-2 govuk-!-margin-bottom-4">
+      <a class="govuk-link" href="/pop/messages">&lt; Back to messages</a>
+      <div class="govuk-grid-row govuk-!-margin-top-6 govuk-!-paddin-bottom-4">
+          <h1 class="govuk-heading-l govuk-!-margin-0 govuk-!-padding-0 govuk-!-padding-bottom-2  ">Write new message</h1>
+          <p class="govuk-body-s govuk-!-color-muted">Send a message to your probation practitioner</p>
+      </div>
+      <div class="govuk-grid-row govuk-!-margin-top-6 ">
+          <div class="govuk-box govuk-box--white govuk-box--rounded govuk-!-padding-6 border box-shadow">
+
+        <form class="">
+          <div class="govuk-form-group">
+            <label class="govuk-label" for="subject">Subject</label>
+            <input class="govuk-input" id="subject" name="subject" type="text" placeholder="">
+          </div>
+          <div class="govuk-form-group">
+            <label class="govuk-label" for="message">Your message</label>
+            <textarea class="govuk-textarea" id="message" name="message" rows="5" placeholder=""></textarea>
+          </div>
+          <div class="govuk-form-group">
+            <label class="govuk-label" for="attachment">Attach supporting evidence (optional)</label>
+            <input class="govuk-file-upload" id="attachment" name="attachment" type="file">
+          </div>
+          <div class="govuk-grid-row">
+            <div class="govuk-grid-column-full govuk-!-text-align-right">
+              {{ govukButton({
+                text: "Cancel",
+                href: "/pop/messages",
+                classes: "govuk-button--secondary govuk-!-margin-right-2 govuk-!-margin-left-2"
+              }) }}
+              {{ govukButton({
+                text: "Send message",
+                type: "submit"
+              }) }}
+            </div>
+          </div>
+        </form>
+        </div>
+      </div>
+    </div>
+  </div>
+</div>
+
+<div class="govuk-grid-row">
+  <div class="govuk-grid-column-full">
+
+  </div>
+</div>
+
+{% endblock %}
