+++ server/views/pages/pop/progress.njk
@@ -1,7 +1,7 @@
 {% extends "../../partials/layout.njk" %}
 {%- from "moj/components/progress-bar/macro.njk" import mojProgressBar -%}
 
-{% set pageTitle = applicationName + " - Home" %}
+{% set pageTitle = applicationName + " - Your progress" %}
 {% set mainClasses = "app-container govuk-body" %}
 
 {% block content %}
